# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventApp API",
        "description": "API REST pour la gestion d'événements collaboratifs - Plateforme de création, gestion et inscription à des événements",
        "version": "1.0.0",
        "contact": {
            "name": "EventApp Support",
            "email": "support@eventapp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "D\u00e9tails d'un \u00e9v\u00e9nement",
                "operationId": "get_api_events_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "D\u00e9tails de l'\u00e9v\u00e9nement"
                    },
                    "404": {
                        "description": "\u00c9v\u00e9nement non trouv\u00e9"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Modifier un \u00e9v\u00e9nement",
                "operationId": "put_api_events_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "startDate",
                                    "location",
                                    "maxParticipants"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Conf\u00e9rence Tech 2024"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une conf\u00e9rence sur les derni\u00e8res technologies"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-15T14:00:00Z"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-15T18:00:00Z"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Paris, France"
                                    },
                                    "maxParticipants": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.99
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "tech",
                                            "conference"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "\u00c9v\u00e9nement modifi\u00e9 avec succ\u00e8s",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Conf\u00e9rence Tech 2024 - Mise \u00e0 jour"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Une conf\u00e9rence sur les derni\u00e8res technologies - Version mise \u00e0 jour"
                                                },
                                                "startDate": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-15T14:00:00Z"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Paris, France"
                                                },
                                                "maxParticipants": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "isPublished": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Donn\u00e9es invalides"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    },
                    "404": {
                        "description": "\u00c9v\u00e9nement non trouv\u00e9"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Supprimer un \u00e9v\u00e9nement",
                "operationId": "delete_api_events_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "\u00c9v\u00e9nement supprim\u00e9"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    },
                    "404": {
                        "description": "\u00c9v\u00e9nement non trouv\u00e9"
                    },
                    "409": {
                        "description": "Impossible de supprimer"
                    }
                }
            }
        },
...
        "/api/registrations": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Mes inscriptions",
                "operationId": "get_api_registrations_list",
                "responses": {
                    "200": {
                        "description": "Liste des inscriptions"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
        "/api/registrations/upcoming": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Mes prochaines inscriptions",
                "operationId": "get_api_registrations_upcoming",
                "responses": {
                    "200": {
                        "description": "Prochaines inscriptions"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
        "/api/registrations/events/{id}": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "S'inscrire \u00e0 un \u00e9v\u00e9nement",
                "operationId": "post_api_registrations_register",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Inscription cr\u00e9\u00e9e"
                    },
                    "400": {
                        "description": "Inscription impossible"
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    },
                    "404": {
                        "description": "\u00c9v\u00e9nement non trouv\u00e9"
                    }
                }
            }
        },
        "/api/registrations/{id}/confirm": {
            "patch": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Confirmer une inscription",
                "operationId": "patch_api_registrations_confirm",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscription confirm\u00e9e"
                    },
                    "400": {
                        "description": "Confirmation impossible"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    },
                    "404": {
                        "description": "Inscription non trouv\u00e9e"
                    }
                }
            }
        },
        "/api/registrations/{id}/cancel": {
            "patch": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Annuler une inscription",
                "operationId": "patch_api_registrations_cancel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscription annul\u00e9e"
                    },
                    "400": {
                        "description": "Annulation impossible"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    },
                    "404": {
                        "description": "Inscription non trouv\u00e9e"
                    }
                }
            }
        },
        "/api/registrations/{id}": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "D\u00e9tails d'une inscription",
                "operationId": "get_api_registrations_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "D\u00e9tails de l'inscription"
                    },
                    "403": {
                        "description": "Acc\u00e8s refus\u00e9"
                    },
                    "404": {
                        "description": "Inscription non trouv\u00e9e"
                    }
                }
            }
        },
        "/api/registrations/ticket/{ticketCode}": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Inscription par code de billet",
                "operationId": "get_api_registrations_by_ticket",
                "parameters": [
                    {
                        "name": "ticketCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscription trouv\u00e9e"
                    },
                    "404": {
                        "description": "Billet non trouv\u00e9"
                    }
                }
            }
        },
...
    },
...
}